REFACT :: [#14] 층별 교실 데이터 분리
